requirements.txt
alembic==1.13.2
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.3
